rev_arrayint

+++ 0x06-pointers_arrays_strings/4-rev_array.c
@@ -1,5 +1,4 @@
 #include "holberton.h"
-
 /**
  * reverse_array - Function that reverses the content of an array of integers.
  * @a: string
@@ -8,7 +7,7 @@
 
 void reverse_array(int *a, int n)
 {
-	short int i, g, aux;
+	int i, g, aux;
 
 	i = 0;
 	for (g = n - 1 ; g  >= (n / 2); g--)
